Update internal API links, de-clutter courses.js
- Give UI components in Courses their own files
  - course-ui.jsx
  - freeblock-ui.jsx
- Add a seats-develop script to develop with local firebase function
- FIX default announcer url in selectionutilities.js
