+++ package.json
@@ -32,6 +32,7 @@
   "scripts": {
     "build": "gatsby build",
     "develop": "gatsby develop",
+    "seats-develop": "GATSBY_COURSES_API=http://localhost:5000/lowellarenarolodex/us-central1/updateSeats gatsby develop",
     "start": "npm run develop",
     "format": "prettier --write \"src/**/*.js\"",
     "test": "echo \"Error: no test specified\" && exit 1"
