+++ src/components/courses.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
-import styles from './courses.module.css';
+import Course from './course-ui';
+import FreeBlock from './freeblock-ui';
 import SelectionUtilities from '../selectionutilities';
 
 /**A container for all the courses from user input.*/
@@ -186,109 +187,5 @@ export default class Courses extends React.Component {
       options: options
     };
     this.setState(state);
-  }
-}
-
-/**An individual Course where the user inputs their class, preferred teacher
-and block, etc.*/
-class Course extends React.Component {
-  removeSelf() {
-    this.props.remove(this.props.id);
-  }
-  render() {
-    return (
-      <div className={styles.course}>
-        <button className={styles.deleteBut}
-          onClick={this.removeSelf.bind(this)}>
-          Remove</button>
-        <CourseSelect name="Subject" parentKey={this.props.id} handleChange={this.props.changeHandler}
-          options={this.props.options}
-          defValue={this.props.default.Subject} />
-        <CourseSelect name="Class" parentKey={this.props.id} handleChange={this.props.changeHandler}
-          options={this.props.options}
-          defValue={this.props.default.Class} />
-        <CourseSelect name="Teacher" parentKey={this.props.id} handleChange={this.props.changeHandler}
-          options={this.props.options}
-          defValue={this.props.default.Teacher} />
-        <CourseSelect name="Block" parentKey={this.props.id} handleChange={this.props.changeHandler}
-          options={this.props.options}
-          defValue={this.props.default.Block} />
-      </div>
-    );
-  }
-}
-
-class CourseSelect extends React.Component {
-  handleChange(e) {
-    this.props.handleChange(this.props.parentKey, this.props.name, e.target.value, false);
-  }
-  handleChangeRequired(e) {
-    this.props.handleChange(this.props.parentKey, this.props.name+'Required', e.target.checked, false);
-  }
-  render() {
-    let checkbox = '';
-    if (this.props.name === 'Teacher')
-      checkbox = (
-        <label style={{textAlign:'right', marginTop:'0.2rem'}}
-          onChange={this.handleChangeRequired.bind(this)}>
-            Is this a required {this.props.name.toLowerCase()}? <input type="checkbox" />
-        </label>
-      );
-    const options = (this.props.options && this.props.options[this.props.name] !== undefined) ?
-      this.props.options[this.props.name] : [];
-    return (
-      <label>
-        {this.props.name}
-        <div>
-          <select defaultValue={this.props.defValue}
-            onChange={this.handleChange.bind(this)}>
-            {options.map((option) =>
-              <option key={option[1]} value={option[1]}>{option[0]}</option>
-            )}
-          </select>
-          {checkbox}
-        </div>
-      </label>
-    );
-  }
-}
-class FreeBlock extends React.Component {
-  removeSelf() {
-    this.props.remove(this.props.id);
-  }
-  handleChange(e) {
-    this.props.changeHandler(this.props.id, e.target.name, e.target.value, true);
-    e.preventDefault();
-  }
-  render() {
-    return (
-      <div className={styles.freeblock}>
-        <button className={styles.deleteBut}
-          onClick={this.removeSelf.bind(this)}>
-          Remove</button>
-        <label>
-          Preferred free block
-          <select onChange={this.handleChange.bind(this)} name="Block"
-            defaultValue={this.props.default.Block}>
-            <option value="">Choose a block</option>
-            <option value="1">1</option>
-            <option value="2">2</option>
-            <option value="3">3</option>
-            <option value="4">4</option>
-            <option value="5">5</option>
-            <option value="6">6</option>
-            <option value="7">7</option>
-            <option value="8">8</option>
-          </select>
-        </label>
-        <label>
-          Free block priority
-          <input
-            value={this.props.default.priorityBlock}
-            type="number" min="1" max="10"
-            onChange={this.handleChange.bind(this)} name="priorityBlock" />
-        </label>
-      </div>
-    );
   }
 }
